+++ apps/app/src/components/views/board-view/board-search-bar.tsx
