+++ apps/app/src/components/views/board-view/components/index.ts
@@ -1,0 +1,2 @@
+export { KanbanCard } from "./kanban-card";
+export { KanbanColumn } from "./kanban-column";
