+++ apps/app/src/components/views/board-view/shared/index.ts
@@ -1,0 +1,5 @@
+export * from "./model-constants";
+export * from "./model-selector";
+export * from "./thinking-level-selector";
+export * from "./profile-quick-select";
+export * from "./testing-tab-content";
