+++ apps/app/src/components/views/profiles-view/utils.ts
@@ -5,3 +5,4 @@ export function getProviderFromModel(model: AgentModel): ModelProvider {
   return "claude";
 }
 
+
